app/requirements.txt
fastapi==0.104.1         # Slight bump for better Pydantic v2 integration
uvicorn[standard]==0.24.0
pydantic==2.6.4          # Latest stable in the 2.6.x line
pydantic-settings==2.2.1
motor==3.7.1             # FIX: Resolves the pymongo import error
pymongo==4.6.3           # Pinning this ensures compatibility with motor 3.4
minio==7.2.0
redis==5.0.1
qdrant-client==1.17.0    # Updated for better reliability
celery==5.3.6
pika==1.3.2
openai==1.14.3           # Updated to resolve common Pydantic v2 "NameErrors"
google-generativeai
python-multipart
pymupdf==1.23.8
tiktoken
fastembed
python-dotenv
passlib[bcrypt]
pyjwt
pydantic[email]
bcrypt==3.1.7